requirements.txt
pillow==11.0.0
python-dotenv==1.0.1
python-telegram-bot==13.15
requests==2.32.3
